pyproject.toml
[tool.poetry]
name = "gaytrifarm"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<4"
django = ">=5.2.7,<6.0.0"
djangorestframework = ">=3.16.1,<4.0.0"
dj-database-url = ">=3.0.1,<4.0.0"
psycopg = ">=3.2.10,<4.0.0"
psycopg2-binary = ">=2.9.10,<3.0.0"
djangorestframework-simplejwt = ">=5.5.1,<6.0.0"
firebase-admin = ">=7.1.0,<8.0.0"
cryptography = ">=46.0.2,<47.0.0"
reportlab = ">=4.4.4,<5.0.0"
celery = ">=5.5.3,<6.0.0"
redis = ">=7.0.1,<8.0.0"
gunicorn = "^23.0.0"
whitenoise = "^6.11.0"
twilio = "^9.8.5"
sib-api-v3-sdk = "^7.6.0"
pillow = "^12.0.0"
django-cleanup = "^9.0.0"
num2words = "^0.5.14"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
